Refactored schemas
