+++ backend/app/main.py
@@ -2,9 +2,9 @@ from contextlib import asynccontextmanager
 
 from fastapi import FastAPI
 
-from app.config import settings
 from app.core.jobs import cleanup_orphaned_jobs
 from app.routes import api_router
+from app.settings import get_settings
 
 
 @asynccontextmanager
@@ -17,6 +17,8 @@ async def lifespan(_app: FastAPI):
     # Shutdown
     print("🔄 Shutting down FastAPI application...")
 
+
+settings = get_settings()
 
 app = FastAPI(title=settings.PROJECT_NAME, root_path="/api", lifespan=lifespan)
